Artwork Parser
* Encapsulated Artwork Parsing into its own class
