+++ LasVegasArtWorkTour/ArtworkVC.swift
@@ -46,35 +46,8 @@ class ArtworkVC: UIViewController {
         
         let response = dataservice.response
         
-        let json = JSON(response)
-        
-        for (_,artRecord):(String, JSON) in json {
-            //print("\(artRecord)")
-            
-            let artName = artRecord["name"].stringValue
-            //print("artName: \(artName)")
-            let artDesc = artRecord["description"].stringValue
-            //print("artDesc: \(artDesc)")
-            let artArtist = artRecord["artist"].stringValue
-            //print("artArtist: \(artArtist)")
-            let artType = artRecord["type"].stringValue
-            //print("artType: \(artType)")
-            let artURL = artRecord["path"].stringValue
-            //print("artURL: \(artURL)")
-            let artGeoLocLat = artRecord["location_1"]["coordinates"][1].stringValue
-            //print("artGeoLoc: \(artGeoLocLat)")
-            let artGeoLocLon = artRecord["location_1"]["coordinates"][0].stringValue
-            //print("artGeoLocLon: \(artGeoLocLon)")
-            
-            let lat = (artGeoLocLat as NSString).doubleValue
-            let lon = (artGeoLocLon as NSString).doubleValue
-            let coordinate = CLLocationCoordinate2D(latitude: lat, longitude: lon)
-            
-            let art = Artwork(name: artName, desc: artDesc, picURL: artURL, artist: artArtist, type: artType, coordinate: coordinate)
-            
-            self.artworks.append(art)
-        }
-        
+        self.artworks = ArtworkParser.parse(response)
+
         print("self.artworks.count: \(self.artworks.count)")
         performSegue(withIdentifier: "ShowArtworkMap", sender: self.artworks)
     }
